+++ Environments/SingleTaxiEnv/single_taxi_wrapper.py
@@ -4,13 +4,13 @@ from Environments.SingleTaxiEnv.single_small_taxi_env import *
 from Environments.abstract_wrapper_env import AbstractWrapperEnv
 
 
-class SingleTaxiSimpleEnv(SingleSmallTaxiEnv, AbstractWrapperEnv):
+class SingleTaxiSimpleEnv(SingleTaxiEnv, AbstractWrapperEnv):
     def __init__(self, deterministic=True):
         super().__init__(deterministic)
         self.transform_num = 7
 
-    def is_possible_initial_state(self, pass_idx, dest_idx, row, col):
-        return pass_idx < 4 and pass_idx != dest_idx and row == self.init_row and col == 0  # Fix taxi initial position
+    def is_possible_initial_state(self, pass_idx, dest_idx, row, col):  # Fix taxi initial position
+        return pass_idx < 4 and pass_idx != dest_idx and row == self.init_row and col == self.init_col
 
     def get_states_from_partial_obs(self, partial_obs):
         partial_obs_aligned_with_env = False
@@ -28,9 +28,9 @@ class SingleTaxiSimpleEnv(SingleSmallTaxiEnv, AbstractWrapperEnv):
                 range(len(self.passengers_locations)))
             passenger_dest = [partial_obs[3]] if (partial_obs[3] is not None) else list(
                 range(len(self.passengers_locations)))
-            fuel = [partial_obs[4]] if partial_obs[4] else list(range(1, MAX_FUEL))
+            fuel = [partial_obs[4]] if partial_obs[4] else list(range(0, MAX_FUEL))
             states = list(product(taxi_x, taxi_y, passenger_loc, passenger_dest, fuel, repeat=1))
-            states = [self.encode(state[0], state[1], state[2], state[3], state[4]) for state in states]
+            states = [self.encode(*state) for state in states]
         else:
             states = []
         return states
